+++ ui/components/src/Table/TableHead.tsx
@@ -1,6 +1,5 @@
-import { TableHead as MuiTableHead, styled, TableHeadProps as MuiTableHeadProps } from '@mui/material';
+import { TableHead as MuiTableHead, TableHeadProps as MuiTableHeadProps } from '@mui/material';
 import { forwardRef } from 'react';
-import { TableDensity } from './table-model';
 
 type TableHeadProps = MuiTableHeadProps;
